requirements.txt

torch
torchvision
ultralytics
supervision>=0.10.0
opencv-python
scikit-learn
numpy
PyYAML
